pyproject.toml
[tool.poetry]
name = "thenewboston-backend"
version = "0.1.0"
description = ""
authors = ["bucky <buckyroberts@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "thenewboston"}]

[tool.poetry.dependencies]
boto3 = "^1.38.23"
channels = {extras = ["daphne"], version = "^4.2.2"}
django = "^5.2.1"
django-cors-headers = "^4.7.0"
django-filter = "^25.1"
django-split-settings = "^1.3.2"
django-storages = "^1.14.6"
djangorestframework = "^3.16.0"
djangorestframework-simplejwt = "^5.5.0"
pillow = "^11.2.1"
psycopg2 = "^2.9.10"
pydantic = "^2.11.5"
pynacl = "^1.5.0"
python = "^3.13"
pyyaml = "^6.0.2"
requests = "^2.32.3"
whitenoise = "^6.9.0"
channels-redis = "^4.2.1"
sentry-sdk = {extras = ["django"], version = "^2.29.1"}
django-restql = "^0.16.2"
django-model-utils = "^5.0.0"
redis = "^6.2.0"

[tool.poetry.group.dev.dependencies]
colorlog = "^6.9.0"
pre-commit = "^4.2.0"
pytest = "^8.3.5"
pytest-xdist = "^3.6.1"
pytest-django = "^4.11.1"
model-bakery = "^1.20.4"
pytest-env = "^1.1.5"
pytest-cov = "^6.1.1"
freezegun = "^1.5.2"
pytest-parametrize-cases = "^0.1.2"
pytest-asyncio = "^0.26.0"
diff-cover = "^9.6.0"

[tool.isort]
line_length = 119
multi_line_output = 5

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "thenewboston.project.settings"
django_find_project = false
python_files = "test_*.py"
testpaths = ["thenewboston"]
filterwarnings = "ignore::DeprecationWarning:^(?!node\\.).*:"
asyncio_default_fixture_loop_scope = "function"
env = [
    "PYTEST_RUNNING = true",
    "THENEWBOSTON_SETTING_SECRET_KEY = dummy",
]

[tool.yapf]
align_closing_bracket_with_visual_indent = true
based_on_style = "google"
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"